pyproject.toml
[build-system]
requires = ["flit_core >=2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pygindex"
authors = [
    {name = "Rytis Sileika", email = "rts.pub@pulegium.eu"},
]
dependencies = [
    "requests",
    "PyYAML"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries",
    "Topic :: Office/Business :: Financial",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
]
readme = "README.md"
dynamic = ["version", "description"]

[project.urls]
documentation = "https://pygindex.readthedocs.io/en/latest/"
source = "https://github.com/rytis/pygindex/"

[project.scripts]
pygi = "pygindex:cli.app"
